+++ Content.Server/_CorvaxNext/Speech/Components/OhioBrainrotAccentComponent.cs
@@ -1,0 +1,8 @@
+using Content.Server._CorvaxNext.Speech.EntitySystems;
+
+namespace Content.Server._CorvaxNext.Speech.Components;
+
+[RegisterComponent]
+[Access(typeof(OhioAccentSystem))]
+public sealed partial class OhioBrainrotAccentComponent : Component
+{ }
