+++ Resources/Locale/en-US/_corvaxnext/accent/accents.ftl
